pyproject.toml
[project]
name = "respeaker2-configurator"
version = "0.1.0"
description = "A simple configurator GUI for the ReSpeaker Mic Array v2.0."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duit[nicegui]>=0.1.17",
    "pyusb>=1.3.1",
    "pywebview>=5.4",
]
